refactor: 파일 위치 수정 & 사용하지 않는 import 삭제
