+++ src/main/java/com/pyro/yolog/domain/member/repository/MemberRepository.java
@@ -4,11 +4,9 @@ import com.pyro.yolog.domain.member.entity.Member;
 import com.pyro.yolog.domain.member.entity.Role;
 import com.pyro.yolog.domain.member.entity.SocialType;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
